app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.30.5
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
vysp-python==0.2.5
Werkzeug==3.0.3
